Refactor checkboxes and radiobuttons that are used in several parts of the application.
