+++ src/app.js
@@ -24,7 +24,7 @@ import { ApplicationInsights } from '@microsoft/applicationinsights-web'
 const environment = process.env.BUILD_ENV
 const dsn = configuration?.sentry?.dsn
 const connectionString = configuration?.azure?.connectionString
-const release = process.env.FRONTEND_TAG
+const release = process.env.FRONTEND_TwAG
 
 if (dsn) {
   Sentry.init({
