+++ src/components/CheckboxList/CheckboxList.test.tsx
@@ -3,12 +3,11 @@
 import { act, render, screen } from '@testing-library/react'
 import 'jest-styled-components'
 import userEvent from '@testing-library/user-event'
-
 import statuses from 'signals/incident-management/definitions/statusList'
 import { withAppContext } from 'test/utils'
 import categories from 'utils/__tests__/fixtures/categories.json'
 
-import CheckboxList from '.'
+import CheckboxList from './index'
 
 describe('signals/incident-management/components/CheckboxList', () => {
   it('should render a title ', () => {
