+++ src/components/CheckboxList/doc/all_checked.png
