+++ src/components/CheckboxList/doc/simple.png
