+++ src/components/CheckboxList/doc/toggle.png
