+++ src/signals/settings/users/Detail/components/UserForm/index.js
@@ -3,10 +3,12 @@
 import { useReducer, useCallback } from 'react'
 
 import { themeSpacing, Row, Column } from '@amsterdam/asc-ui'
+import CheckboxList from 'components/CheckboxList'
 import FormFooter from 'components/FormFooter'
 import History from 'components/History'
 import Input from 'components/Input'
 import Label from 'components/Label'
+import RadioButtonList from 'components/RadioButtonList'
 import TextArea from 'components/TextArea'
 import { makeSelectDepartments } from 'models/departments/selectors'
 import {
@@ -16,8 +18,6 @@ import {
 import PropTypes from 'prop-types'
 import { useSelector } from 'react-redux'
 import { userType, historyType } from 'shared/types'
-import CheckboxList from 'signals/incident-management/components/CheckboxList'
-import RadioButtonList from 'signals/incident-management/components/RadioButtonList'
 import styled from 'styled-components'
 
 const Form = styled.form`
